create login activity and fragments
