+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             android:value="@string/facebook_app_id"/>
 
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -26,6 +26,9 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+
+        <activity android:name=".ui.Login.LoginActivity"></activity>
+
         <activity
             android:name="com.facebook.LoginActivity"
             android:label="@string/app_name"
