--- app/src/main/java/com/yebimom/android/yebimom/Login/LoginActivity.java
@@ -1,10 +1,0 @@
-package com.yebimom.android.yebimom.Login;
-
-import android.app.Activity;
-
-/**
- * Created by corikachu on 15. 3. 7..
- */
-public class LoginActivity extends Activity {
-
-}
